feat(208-03): propagate type narrowing through MintModule, WhaleModule, PayoutUtils
- uint48→uint32 for day-index vars, params, constants, events
- lootboxRng packed reads/writes with scaling helpers in Mint and Whale
- presale packed reads in Mint and Whale
- claimablePool arithmetic cast to uint128

+++ contracts/modules/DegenerusGamePayoutUtils.sol
@@ -99,7 +99,7 @@ abstract contract DegenerusGamePayoutUtils is DegenerusGameStorage {
             unchecked {
                 claimableWinnings[winner] += remainder;
             }
-            claimablePool += remainder;
+            claimablePool += uint128(remainder);
             emit PlayerCredited(winner, winner, remainder);
         }
     }
